+++ src/core/main.rs
@@ -159,6 +159,7 @@ impl Pushrod {
 //        0
 //    }
 
+    /// Retrieves the window `GfxFactory` factory object for graphics textures.
     pub fn get_factory(&mut self) -> &mut GfxFactory {
         &mut self.window.factory
     }